brought up to Java 7 and performance improvements

+++ src/org/saga/factions/FactionDefinition.java
@@ -77,7 +77,7 @@ public class FactionDefinition {
 	 */
 	public Integer getMaxLevel() {
 
-		return levelUpExp.getXMax().intValue();
+		return levelUpExp.getXMax();
 
 	}
